+++ Watchit/src/ContentControl/Content.java
@@ -11,7 +11,7 @@ abstract public class Content {
     public Long contentID;
     public String contentTitle;
     public Date Date;
-    public static Long cnt = (long)1;
+    public static long cnt = (long)1;
     public CastMember[] cast;
     public Director director;
     public String genres;
@@ -24,7 +24,7 @@ abstract public class Content {
     public float Rate_Sum;
 
     public Content(String contentTitle, java.util.Date date, int SizeOfCast, Director director, String genres, String language, String country, int budget, int revenue, Image image) {
-        this.contentID = cnt++;
+        this.contentID = (Long) cnt++;
         this.contentTitle = contentTitle;
         Date = date;
         this.cast = new CastMember[SizeOfCast];
@@ -49,9 +49,9 @@ abstract public class Content {
     }
 
     public void EditRate(int UserID, String contentTitle, float rate){
-        WatchRecord WatchRecordTemp = DataBase.getInstance().watchRecordData.removeData(contentTitle,(long)UserID);
+        WatchRecord WatchRecordTemp = DataBase.getInstance().watchRecordData.removeData(contentTitle, Long.valueOf((long)UserID));
         Rate_Sum -= WatchRecordTemp.Rating;
-        WatchRecordTemp.Rating = rate;
+        WatchRecordTemp.Rating = (Float) rate;
         Rate_Sum += rate;
         DataBase.getInstance().watchRecordData.addData(WatchRecordTemp);
     }
